refactor auth into separate hook, moved delete modal to table component
